server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo[srv]==4.7.2
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0